Add wav_file module that contains WavFile class and test
New module contains WavFile class that represents a wav file read from
disc and contains methods that help validate its meta data. A
WavFileMetaData class is also in the module that contains values of the
meta data at the beginning of file. Also setup test resource under
wav_files dir of test dir that can be passed as the argument to
wav_reader.